backend/requirements.txt
# backend/requirements.txt

# FastAPI and dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7
pydantic==2.5.3

# API integrations
openai==1.6.1
deepgram-sdk==2.12.0
PyJWT==2.8.0
requests==2.31.0

# File processing
PyPDF2==3.0.1
python-dotenv==1.0.0

# Optional - for production
gunicorn==21.2.0